ai知识库检索回答工作流

+++ src/main/java/com/tarzan/maxkb4j/module/dataset/dto/HitTestDTO.java
@@ -7,5 +7,5 @@ public class HitTestDTO {
     private String  query_text;
     private String  search_mode;
     private Double  similarity;
-    private Double  top_number;
+    private Integer  top_number;
 }
